fix: renaming track-event to change-event
